pyproject.toml
[project]
name = "claude-pptx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow==11.0.0",
    "python-pptx==1.0.2",
]
